backend/refactor: moved certain shared functions from vrc to driverOnboarding
