+++ Backend/app/provider-platform/dynamic-offer-driver-app/Main/src/Domain/Action/UI/DriverOnboarding/VehicleRegistrationCertificate.hs
@@ -31,33 +31,11 @@ module Domain.Action.UI.DriverOnboarding.VehicleRegistrationCertificate
     LinkedRC (..),
     DeleteRCReq (..),
     convertTextToUTC,
-    mkIdfyVerificationEntity,
     mkHyperVergeVerificationEntity,
     mkMorthVerificationEntity,
     validateRCResponse,
     VerificationReqRecord (..),
-    DriverDocument (..),
-    checkPan,
-    checkDL,
-    checkAadhaar,
-    validateDocument,
-    compareDateOfBirth,
-    makeDocumentVerificationLockKey,
-    isNameCompareRequired,
-    inferPanTypeFromNumber,
-    isBusinessPan,
-    validateIndividualPANCheck,
-    getDriverDocumentInfo,
-    getDocumentImage,
-    FaceMatchOutcome (..),
-    runDocFaceMatch,
-    runDeferredFaceMatchOnSelfie,
-    isNameComparePercentageValid,
-    imageS3Lock,
-    getRegexRulesFromDocumentConfig,
-    matchesRegexSafely,
     normalizeDocumentNumber,
-    validateByRegex,
   )
 where
 
@@ -68,14 +46,13 @@ import Data.Aeson hiding (Success)
 import qualified Data.HashMap.Strict as HM
 import qualified Data.List as DL
 import qualified Data.Text as T hiding (elem, find, map, zip)
-import Data.Time (Day, utctDay)
+import Data.Time (Day)
 import qualified Domain.Types.Common as DCommon
 import qualified Domain.Types.DocStatus as DocStatus
 import qualified Domain.Types.DocsVerificationStatus as DDVS
 import qualified Domain.Types.DocumentVerificationConfig as ODC
 import qualified Domain.Types.DriverFlowStatus as DDFS
 import qualified Domain.Types.DriverInformation as DI
-import qualified Domain.Types.DriverPanCard as DPan
 import qualified Domain.Types.HyperVergeVerification as Domain
 import qualified Domain.Types.IdfyVerification as Domain
 import qualified Domain.Types.Image as Image
@@ -94,7 +71,6 @@ import Kernel.Beam.Functions
 import Kernel.Beam.Lib.UtilsTH (HasSchemaName)
 import Kernel.External.Encryption
 import Kernel.External.Types (Language (..), SchedulerFlow, ServiceFlow, VerificationFlow)
-import qualified Kernel.External.Verification.Interface as VI
 import qualified Kernel.External.Verification.Types as VT
 import Kernel.Prelude hiding (find)
 import qualified Kernel.Storage.Clickhouse.Config as CH
@@ -121,15 +97,10 @@ import qualified Storage.CachedQueries.Driver.OnBoarding as CQO
 import qualified Storage.CachedQueries.VehicleServiceTier as CQVST
 import Storage.ConfigPilot.Config.DocumentVerificationConfig (DocumentVerificationConfigDimensions (..))
 import Storage.ConfigPilot.Config.TransporterConfig (TransporterConfigDimensions (..))
-import qualified Storage.Queries.AadhaarCard as QAadhaarCard
-import qualified Storage.Queries.DriverGstin as DGQuery
 import qualified Storage.Queries.DriverInformation as DIQuery
-import qualified Storage.Queries.DriverLicense as DLQuery
-import qualified Storage.Queries.DriverPanCard as DPQuery
 import Storage.Queries.DriverRCAssociation (buildRcHM)
 import qualified Storage.Queries.DriverRCAssociation as DAQuery
 import qualified Storage.Queries.FleetDriverAssociationExtra as FDA
-import qualified Storage.Queries.FleetOwnerInformation as FOI
 import qualified Storage.Queries.FleetRCAssociation as FRCAssoc
 import qualified Storage.Queries.HyperVergeVerification as HVQuery
 import qualified Storage.Queries.IdfyVerification as IVQuery
@@ -143,7 +114,6 @@ import qualified Storage.Queries.Vehicle as VQuery
 import qualified Storage.Queries.VehicleDetails as CQVD
 import qualified Storage.Queries.VehicleRegistrationCertificate as RCQuery
 import qualified Storage.Queries.VehicleRegistrationCertificateExtra as VRCExtra
-import Text.Regex.Posix ((=~))
 import Tools.Error
 import qualified Tools.Verification as Verification
 
@@ -206,14 +176,6 @@ data RCValidationReq = RCValidationReq
   }
   deriving (Generic, Show, ToJSON, FromJSON)
 
-data DriverDocument = DriverDocument
-  { panNumber :: Maybe Text,
-    aadhaarNumber :: Maybe Text,
-    dlNumber :: Maybe Text,
-    gstNumber :: Maybe Text
-  }
-  deriving (Generic, Show, ToJSON, FromJSON)
-
 validateDriverRCReq :: Validate DriverRCReq
 validateDriverRCReq DriverRCReq {..} =
   sequenceA_
@@ -230,32 +192,6 @@ prefixMatchedResult rcNumber = DL.any (`T.isPrefixOf` rcNumber)
 normalizeDocumentNumber :: Text -> Text
 normalizeDocumentNumber = T.toUpper . removeSpaceAndDash
 
-getRegexRulesFromDocumentConfig :: ODC.DocumentVerificationConfig -> [Text]
-getRegexRulesFromDocumentConfig config = maybe [] (mapMaybe (.regexValidation)) config.documentFields
-
-matchesRegexSafely :: Text -> Text -> Text -> Flow (Maybe Bool)
-matchesRegexSafely documentType input regexPattern = do
-  result <- try @_ @SomeException $ do
-    let matched = (T.unpack input =~ T.unpack regexPattern :: Bool)
-    matched `seq` pure matched
-  case result of
-    Right matched -> pure (Just matched)
-    Left err -> do
-      logError $ "Invalid regex in DocumentVerificationConfig for " <> documentType <> " validation: " <> regexPattern <> ", error: " <> show err
-      pure Nothing
-
-validateByRegex :: Text -> ODC.DocumentVerificationConfig -> Text -> Flow Bool -> Flow Bool
-validateByRegex documentType config input fallback = do
-  let regexRules = getRegexRulesFromDocumentConfig config
-  if null regexRules
-    then fallback
-    else do
-      regexResults <- mapM (matchesRegexSafely documentType input) regexRules
-      let validRegexResults = mapMaybe (\x -> x) regexResults
-      if null validRegexResults
-        then fallback
-        else pure (or validRegexResults)
-
 isRCNumberFormatValid :: ODC.DocumentVerificationConfig -> Text -> Flow Bool
 isRCNumberFormatValid documentVerificationConfig normalizedRCNumber = do
   let normalizedPrefixList = normalizeDocumentNumber <$> documentVerificationConfig.rcNumberPrefixList
@@ -269,23 +205,6 @@ isRCNumberFormatValid documentVerificationConfig normalizedRCNumber = do
           (null normalizedPrefixList || prefixMatchedResult normalizedRCNumber normalizedPrefixList)
             && isLegacyCertNumFormatValid
   validateByRegex "RC" documentVerificationConfig normalizedRCNumber fallbackCheck
-
--- Define a common function to handle role-based decryption
-getDriverDocumentInfo :: Person.Person -> Flow (Bool, DriverDocument)
-getDriverDocumentInfo person = do
-  case person.role of
-    role | role `elem` [Person.FLEET_OWNER, Person.FLEET_BUSINESS] -> do
-      res <- FOI.findByPrimaryKey person.id >>= fromMaybeM (PersonNotFound person.id.getId)
-      decryptedPanNumber <- mapM decrypt res.panNumber
-      decryptedAadhaarNumber <- mapM decrypt res.aadhaarNumber
-      decryptedGstNumber <- mapM decrypt res.gstNumber
-      return (res.blocked, DriverDocument decryptedPanNumber decryptedAadhaarNumber Nothing decryptedGstNumber)
-    _ -> do
-      res <- DIQuery.findById person.id >>= fromMaybeM (PersonNotFound person.id.getId)
-      decryptedPanNumber <- mapM decrypt res.panNumber
-      decryptedAadhaarNumber <- mapM decrypt res.aadhaarNumber
-      decryptedDlNumber <- mapM decrypt res.dlNumber
-      return (res.blocked, DriverDocument decryptedPanNumber decryptedAadhaarNumber decryptedDlNumber Nothing)
 
 verifyRC ::
   Bool ->
@@ -429,98 +348,6 @@ verifyRC isDashboard mbMerchant (personId, _, merchantOpCityId) req bulkUpload m
     makeVerifyRCHitsCountKey :: Text -> Text
     makeVerifyRCHitsCountKey rcNumber = "VerifyRC:rcNumberHits:" <> rcNumber <> ":hitsCount"
 
-imageS3Lock :: Text -> Text
-imageS3Lock path = "image-s3-lock-" <> path
-
-makeDocumentVerificationLockKey :: Text -> Text
-makeDocumentVerificationLockKey personId = "DocumentVerificationLock:" <> personId
-
-isNameComparePercentageValid :: Id DM.Merchant -> Id DMOC.MerchantOperatingCity -> Verification.NameCompareReq -> Flow Bool
-isNameComparePercentageValid merchantId merchantOpCityId req = do
-  transporterConfig <- getOneConfig (TransporterConfigDimensions {merchantOperatingCityId = merchantOpCityId.getId}) (Just (SCTC.findByMerchantOpCityId merchantOpCityId Nothing)) >>= fromMaybeM (TransporterConfigNotFound merchantOpCityId.getId)
-  case transporterConfig.validNameComparePercentage of
-    Just percentage -> do
-      resp <- Verification.nameCompare merchantId merchantOpCityId req
-      logDebug $ "Name compare percentage response: " <> show resp
-      case resp.nameComparedData of
-        Just percentageData -> return $ percentageData.match_output.name_match >= percentage
-        Nothing -> throwError $ InternalError "Name comparison service returned invalid response"
-    Nothing -> return True -- If percentage not configured, assume valid
-
-getDocumentImage :: Id Person.Person -> Text -> ODC.DocumentType -> Flow Text
-getDocumentImage personId imageId_ expectedDocType = do
-  imageMetadata <- ImageQuery.findById (Id imageId_) >>= fromMaybeM (ImageNotFound imageId_)
-  unless (imageMetadata.verificationStatus == Just Documents.VALID) $
-    throwError (ImageNotValid imageId_)
-  unless (imageMetadata.personId == personId) $
-    throwError (ImageNotFound imageId_)
-  unless (imageMetadata.imageType == expectedDocType) $
-    throwError (ImageInvalidType (show expectedDocType) "")
-  Redis.withLockRedisAndReturnValue (imageS3Lock (imageMetadata.s3Path)) 5 $
-    S3.get $ T.unpack imageMetadata.s3Path
-
--- | Outcome of a server-side selfie-vs-document face match.
---   FMSkip     -> face match not required, or document already SDK-matched on-device.
---   FMPass     -> faces matched.
---   FMFail     -> faces did not match; the document Image has been marked INVALID.
---   FMDeferred -> no source (selfie) image available yet; resolve when the selfie arrives.
-data FaceMatchOutcome = FMSkip | FMPass | FMFail | FMDeferred
-  deriving (Eq, Show)
-
--- | Compare a freshly-uploaded document image against the configured source document
---   (typically the driver's selfie / ProfilePhoto) using the merchant's faceMatchService
---   (IDfy two-image compare). Only runs for non-SDK uploads (workflowTransactionId == Nothing);
---   SDK uploads are already face-matched on-device. On a mismatch the document Image is marked
---   INVALID with FaceMatchFailed so the existing reupload UX surfaces it.
-runDocFaceMatch :: Person.Person -> ODC.DocumentVerificationConfig -> Id Image.Image -> Flow FaceMatchOutcome
-runDocFaceMatch person config docImageId =
-  case config.faceMatchSourceDoc of
-    Nothing -> pure FMSkip
-    Just sourceDocType -> do
-      docImage <- ImageQuery.findById docImageId >>= fromMaybeM (ImageNotFound docImageId.getId)
-      -- Skip when the doc was SDK-matched on-device, or when its image isn't VALID yet
-      -- (e.g. isImageValidationRequired=False leaves it MANUAL_VERIFICATION_REQUIRED) — face match isn't applicable.
-      if isJust docImage.workflowTransactionId || docImage.verificationStatus /= Just Documents.VALID
-        then pure FMSkip
-        else do
-          sourceImages <- ImageQuery.findRecentByPersonIdAndImageType person.id sourceDocType
-          case DL.find ((== Just Documents.VALID) . (.verificationStatus)) sourceImages of
-            Nothing -> pure FMDeferred -- no selfie yet; resolve when it arrives
-            Just sourceImage -> do
-              documentImage1 <- getDocumentImage person.id docImageId.getId config.documentType
-              documentImage2 <- getDocumentImage person.id sourceImage.id.getId sourceDocType
-              resp <-
-                Verification.faceCompare person.merchantId person.merchantOperatingCityId $
-                  VI.FaceCompareReq {documentImage1, documentImage2, driverId = person.id.getId}
-              let matched = (resp.faceComparedData >>= (.is_a_match)) == Just True
-              if matched
-                then pure FMPass
-                else do
-                  ImageQuery.updateVerificationStatusAndFailureReason Documents.INVALID FaceMatchFailed docImageId
-                  logInfo $ "Face match failed for document " <> show config.documentType <> " of driver " <> person.id.getId
-                  pure FMFail
-
--- | When a selfie (ProfilePhoto) is uploaded, retroactively face-match previously-uploaded,
---   non-SDK DL/PAN/Aadhaar documents that bypassed the upload-time gate (no selfie was present then).
---   Image-based (not record-based): we re-check the latest uploaded, image-VALID, non-SDK document image
---   regardless of its 3P record status, so a selfie arriving while 3P is still PENDING is also covered.
---   On a mismatch the document image is marked INVALID (by runDocFaceMatch) and the doc record is flipped
---   INVALID if it already exists; if 3P is still pending, the late webhook's sticky-INVALID guard keeps it
---   INVALID. Only docs whose config has faceMatchSourceDoc set are considered.
-runDeferredFaceMatchOnSelfie :: Person.Person -> Flow ()
-runDeferredFaceMatchOnSelfie person =
-  forM_ [ODC.DriverLicense, ODC.PanCard, ODC.AadhaarCard] $ \docType -> do
-    mbConfig <- listToMaybe <$> CQDVC.findByMerchantOpCityIdAndDocumentType person.merchantOperatingCityId docType Nothing
-    whenJust mbConfig $ \config -> whenJust config.faceMatchSourceDoc $ \_ -> do
-      docImages <- ImageQuery.findRecentByPersonIdAndImageType person.id docType
-      whenJust (DL.find (\img -> img.verificationStatus == Just Documents.VALID && isNothing img.workflowTransactionId) docImages) $ \docImg -> do
-        outcome <- runDocFaceMatch person config docImg.id
-        when (outcome == FMFail) $ case docType of
-          ODC.DriverLicense -> DLQuery.updateVerificationStatus Documents.INVALID docImg.id
-          ODC.PanCard -> DPQuery.updateVerificationStatus Documents.INVALID person.id
-          ODC.AadhaarCard -> QAadhaarCard.updateVerificationStatus Documents.INVALID person.id
-          _ -> pure ()
-
 verifyRCFlow :: Person.Person -> Id DMOC.MerchantOperatingCity -> Bool -> Text -> Id Image.Image -> Maybe UTCTime -> Maybe DVC.VehicleCategory -> Maybe Bool -> Maybe Bool -> Maybe Bool -> EncryptedHashedField 'AsEncrypted Text -> Domain.ImageExtractionValidation -> Maybe Text -> Maybe Text -> Maybe Text -> Flow ()
 verifyRCFlow person merchantOpCityId useCategoryBasedPriority rcNumber imageId dateOfRegistration mbVehicleCategory mbAirConditioned mbOxygen mbVentilator encryptedRC imageExtractionValidation mbUdinNumber mbEngineNumber mbChassisNumber = do
   now <- getCurrentTime
@@ -534,7 +361,7 @@ verifyRCFlow person merchantOpCityId useCategoryBasedPriority rcNumber imageId d
   case verifyRes.verifyRCResp of
     Verification.AsyncResp res -> do
       case res.requestor of
-        VT.Idfy -> IVQuery.create =<< mkIdfyVerificationEntity person res.requestId now imageExtractionValidation encryptedRC dateOfRegistration mbVehicleCategory mbAirConditioned mbOxygen mbVentilator imageId Nothing Nothing
+        VT.Idfy -> IVQuery.create =<< mkRCIdfyVerificationEntity person res.requestId now imageExtractionValidation encryptedRC dateOfRegistration mbVehicleCategory mbAirConditioned mbOxygen mbVentilator imageId Nothing Nothing
         VT.HyperVergeRCDL -> HVQuery.create =<< mkHyperVergeVerificationEntity person res.requestId now imageExtractionValidation encryptedRC dateOfRegistration mbVehicleCategory mbAirConditioned mbOxygen mbVentilator imageId Nothing Nothing res.transactionId
         _ -> throwError $ InternalError ("Service provider not configured to return async responses. Provider Name : " <> (show res.requestor))
       CQO.setVerificationPriorityList person.id verifyRes.remPriorityList
@@ -548,35 +375,6 @@ verifyRCFlow person merchantOpCityId useCategoryBasedPriority rcNumber imageId d
           MorthQuery.create morthEntity
         _ -> logError $ "Handle this case if we want to store the response in the database for provider: " <> show resp.requestor
       void $ onVerifyRC person Nothing resp.response (Just verifyRes.remPriorityList) (Just imageExtractionValidation) (Just encryptedRC) imageId Nothing Nothing (Just resp.requestor) mbVehicleCategory
-
-mkIdfyVerificationEntity :: MonadFlow m => Person.Person -> Text -> UTCTime -> Domain.ImageExtractionValidation -> EncryptedHashedField 'AsEncrypted Text -> Maybe UTCTime -> Maybe DVC.VehicleCategory -> Maybe Bool -> Maybe Bool -> Maybe Bool -> Id Image.Image -> Maybe Int -> Maybe Text -> m Domain.IdfyVerification
-mkIdfyVerificationEntity person requestId now imageExtractionValidation encryptedRC dateOfRegistration mbVehicleCategory mbAirConditioned mbOxygen mbVentilator imageId mbRetryCnt mbStatus = do
-  id <- generateGUID
-  return $
-    Domain.IdfyVerification
-      { id,
-        driverId = person.id,
-        documentImageId1 = imageId,
-        documentImageId2 = Nothing,
-        requestId,
-        docType = ODC.VehicleRegistrationCertificate,
-        documentNumber = encryptedRC,
-        driverDateOfBirth = Nothing,
-        imageExtractionValidation = imageExtractionValidation,
-        issueDateOnDoc = dateOfRegistration,
-        status = fromMaybe "pending" mbStatus,
-        idfyResponse = Nothing,
-        vehicleCategory = mbVehicleCategory,
-        airConditioned = mbAirConditioned,
-        oxygen = mbOxygen,
-        ventilator = mbVentilator,
-        retryCount = Just $ fromMaybe 0 mbRetryCnt,
-        nameOnCard = Nothing,
-        merchantId = Just person.merchantId,
-        merchantOperatingCityId = Just person.merchantOperatingCityId,
-        createdAt = now,
-        updatedAt = now
-      }
 
 mkHyperVergeVerificationEntity :: MonadFlow m => Person.Person -> Text -> UTCTime -> Domain.ImageExtractionValidation -> EncryptedHashedField 'AsEncrypted Text -> Maybe UTCTime -> Maybe DVC.VehicleCategory -> Maybe Bool -> Maybe Bool -> Maybe Bool -> Id Image.Image -> Maybe Int -> Maybe Text -> Maybe Text -> m Domain.HyperVergeVerification
 mkHyperVergeVerificationEntity person requestId now imageExtractionValidation encryptedRC dateOfRegistration mbVehicleCategory mbAirConditioned mbOxygen mbVentilator imageId mbRetryCnt mbStatus transactionId = do
@@ -702,7 +500,7 @@ onVerifyRCHandler person rcVerificationResponse mbVehicleCategory mbAirCondition
               case verifyRes.verifyRCResp of
                 Verification.AsyncResp res -> do
                   case res.requestor of
-                    VT.Idfy -> IVQuery.create =<< mkIdfyVerificationEntity person res.requestId now imageExtractionValidation encryptedRC mbDateOfRegistration mbVehicleCategory mbAirConditioned mbOxygen mbVentilator imageId mbRetryCnt mbReqStatus
+                    VT.Idfy -> IVQuery.create =<< mkRCIdfyVerificationEntity person res.requestId now imageExtractionValidation encryptedRC mbDateOfRegistration mbVehicleCategory mbAirConditioned mbOxygen mbVentilator imageId mbRetryCnt mbReqStatus
                     VT.HyperVergeRCDL -> HVQuery.create =<< mkHyperVergeVerificationEntity person res.requestId now imageExtractionValidation encryptedRC mbDateOfRegistration mbVehicleCategory mbAirConditioned mbOxygen mbVentilator imageId mbRetryCnt mbReqStatus res.transactionId
                     _ -> throwError $ InternalError ("Service provider not configured to return async responses. Provider Name : " <> T.pack (show res.requestor))
                   CQO.setVerificationPriorityList person.id verifyRes.remPriorityList
@@ -1140,153 +938,3 @@ rcVerificationLockKey rcNumber = "VehicleRC::RCNumber-" <> rcNumber
 
 makeFleetOwnerKey :: Text -> Text
 makeFleetOwnerKey vehicleNo = "FleetOwnerId:PersonId-" <> removeSpaceAndDash vehicleNo
-
--- Common function for name comparison
-compareNames :: Id DM.Merchant -> Id DMOC.MerchantOperatingCity -> Maybe Text -> Maybe Text -> Id Person.Person -> Flow Bool
-compareNames merchantId merchantOpCityId mbExtractedName mbVerifiedName personId =
-  case (mbExtractedName, mbVerifiedName) of
-    (Just extractedName, Just verifiedName) -> do
-      let nameCompareReq =
-            Verification.NameCompareReq
-              { extractedName = extractedName,
-                verifiedName = verifiedName,
-                percentage = Just True,
-                driverId = personId.getId
-              }
-      isNameValid <- isNameComparePercentageValid merchantId merchantOpCityId nameCompareReq
-      unless isNameValid $ throwError (MismatchDataError "Name match failed with previously uploaded docs")
-      return True
-    _ -> do
-      logInfo "Name comparison checks not executed."
-      return False
-
-compareDateOfBirth :: Maybe UTCTime -> Maybe UTCTime -> Flow Bool
-compareDateOfBirth mbExtractedValue mbVerifiedValue = do
-  case (mbExtractedValue, mbVerifiedValue) of
-    (Just extractedValue, Just verifiedValue) -> do
-      let extractedDay = utctDay extractedValue
-          verifiedDay = utctDay verifiedValue
-      unless (extractedDay == verifiedDay) $ throwError (MismatchDataError $ "Date of birth mismatch: " <> show extractedDay <> " " <> show verifiedDay)
-      return True
-    _ -> do
-      logInfo "Date of birth checks not executed."
-      return False
-
-checkPan :: Id DM.Merchant -> Id DMOC.MerchantOperatingCity -> Id Person.Person -> Maybe Text -> Maybe UTCTime -> ODC.DocumentType -> Flow Bool
-checkPan merchantId merchantOpCityId personId mbExtractedValue mbDateOfBirthValue verifyingDocumentType = do
-  mdriverPanInformation <- DPQuery.findByDriverId personId
-  case mdriverPanInformation of
-    Just panData -> do
-      if verifyingDocumentType `elem` [ODC.AadhaarCard, ODC.DriverLicense]
-        then validateNameAndDOB merchantId merchantOpCityId mbExtractedValue panData.driverNameOnGovtDB mbDateOfBirthValue panData.driverDob personId
-        else return False
-    Nothing -> throwError $ InternalError "Pan not found"
-
-checkDL :: Id DM.Merchant -> Id DMOC.MerchantOperatingCity -> Id Person.Person -> Maybe Text -> Maybe UTCTime -> ODC.DocumentType -> Flow Bool
-checkDL merchantId merchantOpCityId personId mbExtractedValue mbDateOfBirthValue verifyingDocumentType = do
-  mdriverLicense <- DLQuery.findByDriverId personId
-  case mdriverLicense of
-    Just dlData -> do
-      if verifyingDocumentType `elem` [ODC.AadhaarCard, ODC.PanCard]
-        then validateNameAndDOB merchantId merchantOpCityId mbExtractedValue dlData.driverName mbDateOfBirthValue dlData.driverDob personId
-        else return False
-    Nothing -> throwError $ InternalError "DL not found"
-
-checkAadhaar :: Id DM.Merchant -> Id DMOC.MerchantOperatingCity -> Id Person.Person -> Maybe Text -> Maybe UTCTime -> ODC.DocumentType -> Flow Bool
-checkAadhaar merchantId merchantOpCityId personId mbExtractedValue mbDateOfBirthValue verifyingDocumentType = do
-  aadhaarInfo <- QAadhaarCard.findByPrimaryKey personId
-  case aadhaarInfo of
-    Just aadhaarData -> do
-      if verifyingDocumentType `elem` [ODC.PanCard, ODC.DriverLicense]
-        then validateNameAndDOB merchantId merchantOpCityId mbExtractedValue aadhaarData.nameOnCard mbDateOfBirthValue (parseDateTime =<< aadhaarData.dateOfBirth) personId
-        else return False
-    Nothing -> throwError $ InternalError "Aadhaar not found"
-
-checkGST :: Id Person.Person -> Maybe Text -> Flow ()
-checkGST personId mbPanNumber = do
-  mGstData <- DGQuery.findByDriverId personId
-  case mGstData of
-    Just gstData -> checkTwoPanNumber mbPanNumber gstData.panNumber
-    Nothing -> throwError $ InternalError "GST not found"
-
-validateDocument :: Id DM.Merchant -> Id DMOC.MerchantOperatingCity -> Id Person.Person -> Maybe Text -> Maybe Text -> Maybe Text -> ODC.DocumentType -> DriverDocument -> Flow ()
-validateDocument merchantId merchantOpCityId personId mbNameValue mbDateOfBirthValue mbPanNumber verifyingDocumentType DriverDocument {..} = do
-  let mbUtcDateOfBirthValue = parseDateTime =<< mbDateOfBirthValue
-  let skipPanAadhaarCrossCheck = isBusinessPan panNumber || isBusinessPan mbPanNumber
-  case verifyingDocumentType of
-    ODC.AadhaarCard -> do
-      panChecked <- case panNumber of
-        Just _
-          | not skipPanAadhaarCrossCheck ->
-            checkPan merchantId merchantOpCityId personId mbNameValue mbUtcDateOfBirthValue ODC.AadhaarCard
-        _ -> pure False
-      unless panChecked $
-        whenJust dlNumber $ \_ ->
-          void $ checkDL merchantId merchantOpCityId personId mbNameValue mbUtcDateOfBirthValue ODC.AadhaarCard
-    ODC.PanCard -> do
-      aadhaarChecked <- case aadhaarNumber of
-        Just _
-          | not skipPanAadhaarCrossCheck ->
-            checkAadhaar merchantId merchantOpCityId personId mbNameValue mbUtcDateOfBirthValue ODC.PanCard
-        _ -> pure False
-      unless aadhaarChecked $ do
-        dlChecked <-
-          maybe
-            (pure False)
-            (\_ -> checkDL merchantId merchantOpCityId personId mbNameValue mbUtcDateOfBirthValue ODC.PanCard)
-            dlNumber
-        unless dlChecked $
-          when (isJust gstNumber) $
-            checkGST personId mbPanNumber
-    ODC.DriverLicense -> do
-      aadhaarChecked <-
-        maybe
-          (pure False)
-          (\_ -> checkAadhaar merchantId merchantOpCityId personId mbNameValue mbUtcDateOfBirthValue ODC.DriverLicense)
-          aadhaarNumber
-      unless aadhaarChecked $
-        whenJust panNumber $ \_ ->
-          void $ checkPan merchantId merchantOpCityId personId mbNameValue mbUtcDateOfBirthValue ODC.DriverLicense
-    ODC.GSTCertificate -> checkTwoPanNumber mbPanNumber panNumber
-    _ -> return ()
-
-checkTwoPanNumber :: Maybe Text -> Maybe Text -> Flow ()
-checkTwoPanNumber mbExtractedValue mbVerifiedValue = do
-  case (mbExtractedValue, mbVerifiedValue) of
-    (Just extractedValue, Just verifiedValue) -> do
-      unless (extractedValue == verifiedValue) $ throwError (MismatchDataError "GST not linked with existing PAN")
-      return ()
-    _ -> return ()
-
-validateNameAndDOB :: Id DM.Merchant -> Id DMOC.MerchantOperatingCity -> Maybe Text -> Maybe Text -> Maybe UTCTime -> Maybe UTCTime -> Id Person.Person -> Flow Bool
-validateNameAndDOB merchantId merchantOpCityId mbExtractedName mbVerifiedName mbExtractedDOB mbVerifiedDOB personId = do
-  isNameValid <- compareNames merchantId merchantOpCityId mbExtractedName mbVerifiedName personId
-  isDateOfBirthValid <- compareDateOfBirth mbExtractedDOB mbVerifiedDOB
-  return (isNameValid && isDateOfBirthValid)
-
--- | Returns True if name compare is required for the given transporterConfig and verifyBy
-isNameCompareRequired :: DTC.TransporterConfig -> DPan.VerifiedBy -> Bool
-isNameCompareRequired transporterConfig verifyBy =
-  isJust transporterConfig.validNameComparePercentage && verifyBy /= DPan.DASHBOARD_ADMIN && verifyBy /= DPan.DASHBOARD_USER
-
--- (P = Individual, C = Company, H = HUF, F = Firm, A = AOP, T = Trust, B = BOI, L/J/G = others).
-inferPanTypeFromNumber :: Text -> Maybe DPan.PanType
-inferPanTypeFromNumber panNumber =
-  let upperPan = T.toUpper panNumber
-   in if T.length upperPan >= 4
-        then Just $ if T.index upperPan 3 == 'P' then DPan.INDIVIDUAL else DPan.BUSINESS
-        else Nothing
-
-isBusinessPan :: Maybe Text -> Bool
-isBusinessPan = maybe False ((== Just DPan.BUSINESS) . inferPanTypeFromNumber)
-
--- | When a merchant enables individualPANCheck, reject business PANs (4th char ≠ 'P')
--- for individual drivers (incl. fleet drivers) and individual fleet owners.
--- Business fleet owners and other roles are exempt, as they may hold a business PAN.
-validateIndividualPANCheck :: DTC.TransporterConfig -> Person.Person -> Text -> Flow ()
-validateIndividualPANCheck transporterConfig person panNumber =
-  when (transporterConfig.individualPANCheck == Just True && isIndividualRole person.role) $
-    when (inferPanTypeFromNumber (removeSpaceAndDash panNumber) /= Just DPan.INDIVIDUAL) $
-      throwError (InvalidRequest "Business PAN card not be accepted please upload individual PAN")
-  where
-    isIndividualRole role = role == Person.DRIVER || role == Person.FLEET_OWNER
